pyproject.toml
[tool.poetry]
name = "triptailor"
version = "0.1.0"
description = ""
authors = ["Aleksandr Shishkov"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12"
loguru = "^0.7.2"
python-dotenv = "^1.0.1"
langchain = "^0.3.1"
langgraph = "^0.2.34"
langchain-openai = "^0.2.1"
langgraph-checkpoint-sqlite = "^2.0.0"
ipython = "^8.28.0"
langchain-community = "^0.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
